+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Fixed compilation errors caused by late `windows.h` include
 - Fixed ESP not being cleared when not holding down the aim key
 - Fixed vsync not working
+- Fixed crash when no models are in the `models` directory